requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
chardet==3.0.4
discord.py==1.0.0a1426+g8de382f
idna==2.6
multidict==4.0.0
websockets==3.4
yarl==1.0.0
